Move Artefact to content package
Artefact will be rendered by the existing `content_api` package and the
forthcoming `content_store` package. This commit moves it to a more
conceptually general package.
