+++ content_api/artefact.go
@@ -3,28 +3,9 @@ package content_api
 import (
 	"encoding/json"
 
+	. "github.com/alphagov/metadata-api/content"
 	"github.com/alphagov/metadata-api/request"
 )
-
-type Part struct {
-	WebURL  string `json:"web_url"`
-	Title   string `json:"title"`
-}
-
-type Detail struct {
-	NeedIDs             []string `json:"need_ids"`
-	BusinessProposition bool     `json:"business_proposition"`
-	Description         string   `json:"description"`
-	Parts               []Part   `json:"parts"`
-}
-
-type Artefact struct {
-	ID      string `json:"id"`
-	WebURL  string `json:"web_url"`
-	Title   string `json:"title"`
-	Format  string `json:"format"`
-	Details Detail `json:"details"`
-}
 
 func FetchArtefact(contentAPI, bearerToken, slug string) (*Artefact, error) {
 	artefactResponse, err := request.NewRequest(contentAPI+slug+".json", bearerToken)
